requirements.txt
pandas==2.1.4
numpy==1.26.3
jsonschema==4.19.2
pytest==7.4.3
apache-airflow==2.10.5
flake8==6.1.0
black==23.12.1
pre-commit==3.6.2
pytest-mock==3.14.0
python-dotenv